refactor(next): move+rename NextAppPathStrategyPathCanonicalizer → PathCanonicalizer in core
Extract the canonicalizer into src/core/path-canonicalizer.ts as PathCanonicalizer,
export it from the core barrel, and consume it from the strategy via #r-machine/next/core.
Its tests move to dedicated tests/core/path-canonicalizer.test{,-d}.ts, matching the
one-file-one-test convention; the strategy tests keep only the instance-of check.
