+++ packages/r-machine-next/src/core/index.ts
@@ -25,4 +25,5 @@ export {
   type PathAtlasClass,
   type Segment,
 } from "./path-atlas.js";
+export { PathCanonicalizer } from "./path-canonicalizer.js";
 export type { RMachineProxy } from "./proxy.js";
